requirement.txt
streamlit
watchdog
pandas
openpyxl
google-genai
pymupdf
colorama